+++ lib/python/bins.cpp
@@ -148,6 +148,14 @@ void init_buckets(py::module &m) {
                             to_string(dt));
   });
 
+  m.def(
+      "lookup_previous",
+      [](const DataArray &function, const Variable &x, const std::string &dim,
+         const std::optional<Variable> &fill_value) {
+        return dataset::lookup_previous(function, x, Dim{dim}, fill_value);
+      },
+      py::call_guard<py::gil_scoped_release>());
+
   auto buckets = m.def_submodule("buckets");
   buckets.def(
       "concatenate",
@@ -187,8 +195,9 @@ void init_buckets(py::module &m) {
       py::call_guard<py::gil_scoped_release>());
   buckets.def(
       "map",
-      [](const DataArray &function, const Variable &x, const std::string &dim) {
-        return dataset::buckets::map(function, x, Dim{dim});
+      [](const DataArray &function, const Variable &x, const std::string &dim,
+         const std::optional<Variable> &fill_value) {
+        return dataset::buckets::map(function, x, Dim{dim}, fill_value);
       },
       py::call_guard<py::gil_scoped_release>());
   buckets.def(