+++ lib/python/unit.cpp
@@ -27,8 +27,8 @@ get_time_unit(const std::optional<scipp::units::Unit> value_unit,
               const std::optional<scipp::units::Unit> dtype_unit,
               const units::Unit sc_unit) {
   if (!temporal_or_dimensionless(sc_unit)) {
-    throw std::invalid_argument("Invalid unit for dtype=datetime64: " +
-                                to_string(sc_unit));
+    throw except::UnitError("Invalid unit for dtype=datetime64: " +
+                            to_string(sc_unit));
   }
   if (dtype_unit.value_or(units::one) != units::one &&
       (sc_unit != units::one && *dtype_unit != sc_unit)) {
@@ -70,8 +70,8 @@ std::tuple<scipp::units::Unit, scipp::units::Unit>
 common_unit<scipp::core::time_point>(const pybind11::object &values,
                                      const scipp::units::Unit unit) {
   if (!temporal_or_dimensionless(unit)) {
-    throw std::invalid_argument("Invalid unit for dtype=datetime64: " +
-                                to_string(unit));
+    throw except::UnitError("Invalid unit for dtype=datetime64: " +
+                            to_string(unit));
   }
 
   if (values.is_none() || !has_datetime_dtype(values)) {
@@ -87,9 +87,25 @@ common_unit<scipp::core::time_point>(const pybind11::object &values,
 }
 
 std::string to_numpy_time_string(const scipp::units::Unit unit) {
+  if (unit == units::m) {
+    // Would be treated as minute otherwise.
+    throw except::UnitError("Invalid time unit, got 'm' which means meter. "
+                            "If you meant minute, use unit='min' instead.");
+  }
   return unit == units::us            ? std::string("us")
          : unit == units::Unit("min") ? std::string("m")
                                       : to_string(unit);
+}
+
+std::string to_numpy_time_string(const ProtoUnit &unit) {
+  return std::visit(
+      overloaded{
+          [](const scipp::units::Unit &u) { return to_numpy_time_string(u); },
+          [](const std::string &u) {
+            return to_numpy_time_string(scipp::units::Unit(u));
+          },
+          [](const auto &) { return std::string(); }},
+      unit);
 }
 
 scipp::units::Unit unit_or_default(const ProtoUnit &unit, const DType type) {
@@ -100,7 +116,7 @@ scipp::units::Unit unit_or_default(const ProtoUnit &unit, const DType type) {
                          "Default unit requested but dtype unknown.");
                    return variable::default_unit_for(type);
                  },
-                 [](py::none) { return units::none; },
+                 [](const py::none &) { return units::none; },
                  [](const std::string &u) { return units::Unit(u); },
                  [](const units::Unit &u) { return u; }},
       unit);