+++ lib/python/variable.cpp
@@ -18,7 +18,6 @@
 #include "scipp/variable/variable_factory.h"
 
 #include "scipp/dataset/dataset.h"
-#include "scipp/dataset/util.h"
 
 #include "bind_data_access.h"
 #include "bind_operators.h"
@@ -89,17 +88,8 @@ Array of values with dimension labels and a unit, optionally including an array
 of variances.)");
 
   bind_init(variable);
-  variable
-      .def("rename_dims", &rename_dims<Variable>, py::arg("dims_dict"),
-           py::pos_only(), "Rename dimensions.")
-      .def_property_readonly("dtype", &Variable::dtype)
-      .def("__sizeof__",
-           [](const Variable &self) {
-             return size_of(self, SizeofTag::ViewOnly);
-           })
-      .def("underlying_size", [](const Variable &self) {
-        return size_of(self, SizeofTag::Underlying);
-      });
+  variable.def("_rename_dims", &rename_dims<Variable>)
+      .def_property_readonly("dtype", &Variable::dtype);
 
   bind_common_operators(variable);