+++ src/scipp/__init__.py
@@ -49,6 +49,11 @@ setattr(DataArray, 'to_hdf5', _to_hdf5)
 setattr(Dataset, 'to_hdf5', _to_hdf5)
 del _to_hdf5
 
+from .format import format_variable as _format_variable
+
+setattr(Variable, '__format__', _format_variable)
+del _format_variable
+
 from .extend_units import *
 from .object_list import _repr_html_
 from .utils import collapse, slices
@@ -107,6 +112,7 @@ _binding.bind_functions_as_methods(Dataset, globals(), ('squeeze', ))
 for _cls in (DataArray, Dataset):
     _binding.bind_functions_as_methods(_cls, globals(), ('groupby', 'transform_coords'))
 del _cls
+_binding.bind_functions_as_methods(Variable, globals(), ('bin', 'hist', 'nanhist'))
 _binding.bind_functions_as_methods(DataArray, globals(),
                                    ('bin', 'group', 'hist', 'nanhist', 'rebin'))
 _binding.bind_functions_as_methods(Dataset, globals(), ('hist', 'rebin'))