+++ src/scipp/experimental/plotting/line.py
@@ -1,9 +1,9 @@
 # SPDX-License-Identifier: BSD-3-Clause
 # Copyright (c) 2022 Scipp contributors (https://github.com/scipp)
 
-from ... import DataArray, log10, stddevs
+from ... import DataArray, stddevs
 from .tools import get_line_param
-from .limits import find_limits, fix_empty_range
+from .limits import find_limits, fix_empty_range, delta
 
 from functools import reduce
 import numpy as np
@@ -35,6 +35,7 @@ class Line:
         self._errorbars = errorbars
         self._dim = None
         self._unit = None
+        self.label = None
 
         self._mask_color = mask_color if mask_color is not None else 'k'
 
@@ -57,7 +58,7 @@ class Line:
         }
 
     def _make_line(self, data, errorbars, params):
-        label = data["name"]
+        self.label = data["name"]
         has_mask = data["mask"] is not None
         mask_data_key = "mask" if has_mask else "values"
 
@@ -65,7 +66,7 @@ class Line:
             self._line = self._ax.step(
                 data["values"]["x"],
                 data["values"]["y"],
-                label=label,
+                label=self.label,
                 zorder=10,
                 **{key: params[key]
                    for key in ["color", "linewidth"]})[0]
@@ -79,7 +80,7 @@ class Line:
         else:
             self._line = self._ax.plot(data["values"]["x"],
                                        data["values"]["y"],
-                                       label=label,
+                                       label=self.label,
                                        zorder=10,
                                        **params)[0]
             self._mask = self._ax.plot(data["values"]["x"],
@@ -173,21 +174,19 @@ class Line:
         ymin, ymax = fix_empty_range(find_limits(self._data.data, scale=yscale))
 
         # Add padding
+        deltax = delta(xmin, xmax, 0.03, xscale)
         if xscale == "log":
-            delta = 10**(0.03 * log10(xmax / xmin))
-            xmin /= delta
-            xmax *= delta
+            xmin = xmin / deltax
+            xmax = xmax * deltax
         else:
-            delta = 0.03 * (xmax - xmin)
-            xmin -= delta
-            xmax += delta
+            xmin = xmin - deltax
+            xmax = xmax + deltax
+        deltay = delta(ymin, ymax, 0.03, yscale)
         if yscale == "log":
-            delta = 10**(0.03 * log10(ymax / ymin))
-            ymin /= delta
-            ymax *= delta
+            ymin = ymin / deltay
+            ymax = ymax * deltay
         else:
-            delta = 0.03 * (ymax - ymin)
-            ymin -= delta
-            ymax += delta
+            ymin = ymin - deltay
+            ymax = ymax + deltay
 
         return xmin, xmax, ymin, ymax